Refactor third parties frontend to page arborescence
Mirror the risks refactor (c78a713): colocate routes.ts, split the
detail layout query so each child route owns its Loader + Page, rename
tabs/*Tab to resource folders with *Page, move dialogs into
_components/, and extract ThirdPartyRow with its own fragment.

Remove outlet context data passing, deprecated loaderFromQueryLoader,
and the monolithic ThirdPartyGraph node/list queries. List and layout
queries now live in ThirdPartiesPage and ThirdPartyDetailLayout.
