+++ apps/console/src/pages/organizations/third-parties/_components/DeleteDataPrivacyAgreementDialog.tsx
