+++ apps/console/src/pages/organizations/third-parties/_components/EditServiceDialog.tsx
@@ -43,7 +43,7 @@ const updateServiceMutation = graphql`
   mutation EditServiceDialogUpdateMutation($input: UpdateThirdPartyServiceInput!) {
     updateThirdPartyService(input: $input) {
       thirdPartyService {
-        ...ThirdPartyServicesTabFragment_service
+        ...ThirdPartyServiceRow_service
       }
     }
   }
