+++ apps/console/src/pages/organizations/third-parties/_components/VettingDialog.tsx
@@ -44,8 +44,8 @@ const vetMutation = graphql`
         websiteUrl
         vettingStatus
         ...useThirdPartyFormFragment
-        ...ThirdPartyComplianceTabFragment
-        ...ThirdPartyRiskAssessmentTabFragment
+        ...ThirdPartyCompliancePageFragment
+        ...ThirdPartyRiskAssessmentPageFragment
       }
     }
   }
