+++ apps/console/src/pages/organizations/third-parties/risks/_components/ThirdPartyRiskAssessmentRow.tsx
@@ -1,0 +1,96 @@
+// Copyright (c) 2025-2026 Probo Inc <hello@probo.com>.
+//
+// Permission to use, copy, modify, and/or distribute this software for any
+// purpose with or without fee is hereby granted, provided that the above
+// copyright notice and this permission notice appear in all copies.
+//
+// THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
+// REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
+// AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
+// INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
+// LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
+// OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
+// PERFORMANCE OF THIS SOFTWARE.
+
+import { formatDate } from "@probo/helpers";
+import { useTranslate } from "@probo/i18n";
+import {
+  Badge,
+  RiskBadge,
+  Td,
+  Tr,
+} from "@probo/ui";
+import { clsx } from "clsx";
+import { graphql, useFragment } from "react-relay";
+
+import type { ThirdPartyRiskAssessmentRow_assessment$key } from "#/__generated__/core/ThirdPartyRiskAssessmentRow_assessment.graphql";
+
+const riskAssessmentRowFragment = graphql`
+  fragment ThirdPartyRiskAssessmentRow_assessment on ThirdPartyRiskAssessment {
+    id
+    createdAt
+    expiresAt
+    dataSensitivity
+    businessImpact
+    notes
+  }
+`;
+
+interface ThirdPartyRiskAssessmentRowProps {
+  assessmentKey: ThirdPartyRiskAssessmentRow_assessment$key;
+  isExpanded: boolean;
+  onClick: (id: string) => void;
+}
+
+export function ThirdPartyRiskAssessmentRow(props: ThirdPartyRiskAssessmentRowProps) {
+  const { __ } = useTranslate();
+  const assessment = useFragment(riskAssessmentRowFragment, props.assessmentKey);
+  const { relativeDateFormat } = useTranslate();
+  const isExpired = new Date(assessment.expiresAt) < new Date();
+
+  return (
+    <>
+      <Tr
+        className={clsx(
+          isExpired && "opacity-50",
+          "cursor-pointer",
+          props.isExpanded && "border-none",
+        )}
+        onClick={() => props.onClick(assessment.id)}
+      >
+        <Td>
+          <span className="text-xs text-txt-secondary ml-1">
+            {formatDate(assessment.createdAt)}
+          </span>
+        </Td>
+        <Td>
+          <div className="flex items-center gap-2">
+            {relativeDateFormat(assessment.expiresAt)}
+            {isExpired && <Badge variant="neutral">{__("Expired")}</Badge>}
+          </div>
+        </Td>
+        <Td>
+          <RiskBadge level={assessment.dataSensitivity} />
+        </Td>
+        <Td>
+          <RiskBadge level={assessment.businessImpact} />
+        </Td>
+      </Tr>
+      {props.isExpanded && (
+        <Tr className={clsx("border-none", isExpired && "opacity-50")}>
+          <Td colSpan={4}>
+            <div className="space-y-2">
+              <div>
+                {__("Notes")}
+                :
+              </div>
+              <p className="text-sm text-txt-secondary whitespace-pre-wrap">
+                {assessment.notes}
+              </p>
+            </div>
+          </Td>
+        </Tr>
+      )}
+    </>
+  );
+}
