--- apps/console/src/pages/organizations/third-parties/tabs/ThirdPartyServicesTab.tsx
@@ -1,237 +1,0 @@
-// Copyright (c) 2025-2026 Probo Inc <hello@probo.com>.
-//
-// Permission to use, copy, modify, and/or distribute this software for any
-// purpose with or without fee is hereby granted, provided that the above
-// copyright notice and this permission notice appear in all copies.
-//
-// THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
-// REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
-// AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
-// INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
-// LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
-// OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
-// PERFORMANCE OF THIS SOFTWARE.
-
-import { sprintf } from "@probo/helpers";
-import { usePageTitle } from "@probo/hooks";
-import { useTranslate } from "@probo/i18n";
-import {
-  ActionDropdown,
-  Button,
-  DropdownItem,
-  IconPencil,
-  IconPlusLarge,
-  IconTrashCan,
-  PageHeader,
-  Tbody,
-  Td,
-  Th,
-  Thead,
-  Tr,
-  useConfirm,
-} from "@probo/ui";
-import { type ComponentProps, useState } from "react";
-import { useFragment, useRefetchableFragment } from "react-relay";
-import { useOutletContext } from "react-router";
-import { graphql } from "relay-runtime";
-
-import type { ThirdPartyGraphNodeQuery$data } from "#/__generated__/core/ThirdPartyGraphNodeQuery.graphql";
-import type { ThirdPartyServicesListQuery } from "#/__generated__/core/ThirdPartyServicesListQuery.graphql";
-import type { ThirdPartyServicesTabFragment$key } from "#/__generated__/core/ThirdPartyServicesTabFragment.graphql";
-import type {
-  ThirdPartyServicesTabFragment_service$data,
-  ThirdPartyServicesTabFragment_service$key,
-} from "#/__generated__/core/ThirdPartyServicesTabFragment_service.graphql";
-import { SortableTable, SortableTh } from "#/components/SortableTable";
-import { useMutationWithToasts } from "#/hooks/useMutationWithToasts";
-
-import { CreateServiceDialog } from "../dialogs/CreateServiceDialog";
-import { EditServiceDialog } from "../dialogs/EditServiceDialog";
-
-export const thirdPartyServicesFragment = graphql`
-  fragment ThirdPartyServicesTabFragment on ThirdParty
-  @refetchable(queryName: "ThirdPartyServicesListQuery")
-  @argumentDefinitions(
-    first: { type: "Int", defaultValue: 50 }
-    order: { type: "ThirdPartyServiceOrder", defaultValue: null }
-    after: { type: "CursorKey", defaultValue: null }
-    before: { type: "CursorKey", defaultValue: null }
-    last: { type: "Int", defaultValue: null }
-  ) {
-    services(
-      first: $first
-      after: $after
-      last: $last
-      before: $before
-      orderBy: $order
-    ) @connection(key: "ThirdPartyServicesTabFragment_services") {
-      __id
-      edges {
-        node {
-          id
-          canUpdate: permission(action: "core:thirdParty-service:update")
-          canDelete: permission(action: "core:thirdParty-service:delete")
-          ...ThirdPartyServicesTabFragment_service
-        }
-      }
-    }
-  }
-`;
-
-const serviceFragment = graphql`
-  fragment ThirdPartyServicesTabFragment_service on ThirdPartyService {
-    id
-    name
-    description
-    canUpdate: permission(action: "core:thirdParty-service:update")
-    canDelete: permission(action: "core:thirdParty-service:delete")
-  }
-`;
-
-const deleteServiceMutation = graphql`
-  mutation ThirdPartyServicesTabDeleteServiceMutation(
-    $input: DeleteThirdPartyServiceInput!
-    $connections: [ID!]!
-  ) {
-    deleteThirdPartyService(input: $input) {
-      deletedThirdPartyServiceId @deleteEdge(connections: $connections)
-    }
-  }
-`;
-
-export default function ThirdPartyServicesTab() {
-  const { thirdParty } = useOutletContext<{
-    thirdParty: ThirdPartyGraphNodeQuery$data["node"];
-  }>();
-  const [data, refetch] = useRefetchableFragment<
-    ThirdPartyServicesListQuery,
-    ThirdPartyServicesTabFragment$key
-  >(thirdPartyServicesFragment, thirdParty);
-  const connectionId = data.services.__id;
-  const services = data.services.edges.map(edge => edge.node);
-  const { __ } = useTranslate();
-  const [editingService, setEditingService]
-    = useState<ThirdPartyServicesTabFragment_service$data | null>(null);
-  const hasAnyAction = services.some(
-    ({ canUpdate, canDelete }) => canUpdate || canDelete,
-  );
-
-  usePageTitle(thirdParty.name + " - " + __("Services"));
-
-  return (
-    <div className="space-y-6">
-      <PageHeader
-        title={__("Services")}
-        description={__("Manage services provided by this third party.")}
-      >
-        {thirdParty.canCreateService && (
-          <CreateServiceDialog thirdPartyId={thirdParty.id} connectionId={connectionId}>
-            <Button icon={IconPlusLarge}>{__("Add service")}</Button>
-          </CreateServiceDialog>
-        )}
-      </PageHeader>
-
-      <SortableTable
-        refetch={refetch as ComponentProps<typeof SortableTable>["refetch"]}
-      >
-        <Thead>
-          <Tr>
-            <SortableTh field="NAME">{__("Name")}</SortableTh>
-            <Th>{__("Description")}</Th>
-            {hasAnyAction && <Th>{__("Actions")}</Th>}
-          </Tr>
-        </Thead>
-        <Tbody>
-          {services.map(service => (
-            <ServiceRow
-              key={service.id}
-              serviceKey={service}
-              connectionId={connectionId}
-              onEdit={setEditingService}
-            />
-          ))}
-        </Tbody>
-      </SortableTable>
-
-      {editingService && editingService.canUpdate && (
-        <EditServiceDialog
-          serviceId={editingService.id}
-          service={editingService}
-          onClose={() => setEditingService(null)}
-        />
-      )}
-    </div>
-  );
-}
-
-type ServiceRowProps = {
-  serviceKey: ThirdPartyServicesTabFragment_service$key;
-  connectionId: string;
-  onEdit: (service: ThirdPartyServicesTabFragment_service$data) => void;
-};
-
-function ServiceRow(props: ServiceRowProps) {
-  const { __ } = useTranslate();
-  const service = useFragment<ThirdPartyServicesTabFragment_service$key>(
-    serviceFragment,
-    props.serviceKey,
-  );
-  const confirm = useConfirm();
-  const [deleteService] = useMutationWithToasts(deleteServiceMutation, {
-    successMessage: __("Service deleted successfully"),
-    errorMessage: __("Failed to delete service"),
-  });
-  const hasAnyAction = service.canUpdate || service.canDelete;
-
-  const handleDelete = () => {
-    confirm(
-      () =>
-        deleteService({
-          variables: {
-            connections: [props.connectionId],
-            input: {
-              thirdPartyServiceId: service.id,
-            },
-          },
-        }),
-      {
-        message: sprintf(
-          __(
-            "This will permanently delete the service \"%s\". This action cannot be undone.",
-          ),
-          service.name,
-        ),
-      },
-    );
-  };
-
-  return (
-    <Tr>
-      <Td>{service.name}</Td>
-      <Td>{service.description || __("—")}</Td>
-      {hasAnyAction && (
-        <Td width={50} className="text-end">
-          <ActionDropdown>
-            {service.canUpdate && (
-              <DropdownItem
-                icon={IconPencil}
-                onClick={() => props.onEdit(service)}
-              >
-                {__("Edit")}
-              </DropdownItem>
-            )}
-            {service.canDelete && (
-              <DropdownItem
-                icon={IconTrashCan}
-                onClick={handleDelete}
-                variant="danger"
-              >
-                {__("Delete")}
-              </DropdownItem>
-            )}
-          </ActionDropdown>
-        </Td>
-      )}
-    </Tr>
-  );
-}
